SmartVender-Client/requirements.txt
# RPI RPi.GPIO
signalrcore
python-json-logger
requests
